backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0

celery==5.4.0
redis==5.0.8
kombu>=5.3.4

PyYAML==6.0.2

python-docx==1.1.2
lxml>=5.2,<6.0

# 数据库（推荐二选一）
pymysql==1.1.2
# mysqlclient>=2.2.4   # ← 更稳（macOS 强烈推荐）

python-dotenv>=1.0.1

requests==2.31.0
openai==1.3.5
